feat(promo): move copy to own promo mechanism
