+++ internal/controller/promotion/kustomize_test.go
@@ -172,6 +172,7 @@ func TestKustomizerApply(t *testing.T) {
 				"",
 				"",
 				"",
+				nil,
 				git.RepoCredentials{},
 			)
 			testCase.assertions(t, changes, err)
